BOL-122 feat: user rank UiModel 변경에 따른 adapter 관련 코드 수정
