+++ presentation/src/main/java/com/yapp/bol/presentation/model/UserRankUiModel.kt
@@ -3,6 +3,6 @@ package com.yapp.bol.presentation.model
 import com.yapp.bol.domain.model.UserRankItem
 
 sealed class UserRankUiModel {
-    data class UserRank1to3(val userRankItem: UserRankItem) : UserRankUiModel()
+    data class UserRank1to3(val userRankItemList: List<UserRankItem>) : UserRankUiModel()
     data class UserRankAfter4(val userRankItem: UserRankItem) : UserRankUiModel()
 }
